python/requirements-docker.txt
# Generated by dev.py on Tue Nov  5 17:54:59 2024
# This File excludes Windows-specific Python libraries
# from the Docker image.

aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
azure-core==1.31.0
azure-identity==1.19.0
azure-storage-blob==12.23.1
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
coverage[toml]==7.6.4
cryptography==43.0.3
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
faker==26.1.0
fastapi==0.115.3
frozenlist==1.5.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.10.0
markupsafe==3.0.2
more-itertools==10.5.0
motor==3.6.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.52.2
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==24.1
parse==1.20.2
pathable==0.4.3
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
prance==23.6.21.0
priority==2.0.0
propcache==0.2.0
psycopg[binary,pool]==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pyjwt[crypto]==2.9.0
pymongo==4.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.15
pyyaml==6.0.2
referencing==0.30.2
regex==2023.12.25
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.20.0
ruamel-yaml==0.18.6
semantic-kernel==0.9.1b1
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
tiktoken==0.8.0
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
werkzeug==3.0.6
wsproto==1.2.0
xmlformatter==0.2.6
yarl==1.16.0
